Allow waiting for successful start of async commands (#67)
Adds a Builder for configuring CloudSdk, and manages DefaultProcessRunnner configuration for the clients by exposing settings in the Builder.

+++ src/main/java/com/google/cloud/tools/app/impl/cloudsdk/CloudSdkAppEngineStandardStaging.java
@@ -74,7 +74,7 @@ public class CloudSdkAppEngineStandardStaging implements AppEngineStandardStagin
 
       cloudSdk.runAppCfgCommand(arguments);
 
-      if (dockerfile != null) {
+      if (dockerfile != null && dockerfile.toFile().exists()) {
         Files.copy(dockerfile, dockerfileDestination, StandardCopyOption.REPLACE_EXISTING);
       }
